+++ enterprise/frontend/src/embedding-sdk/components/public/MetabaseProvider.tsx
@@ -19,8 +19,8 @@ import type {
   SdkStoreState,
 } from "embedding-sdk/store/types";
 import type { MetabaseAuthConfig } from "embedding-sdk/types";
-import type { MetabaseTheme } from "embedding-sdk/types/theme";
 import { EMBEDDING_SDK_ROOT_ELEMENT_ID } from "metabase/embedding-sdk/config";
+import type { MetabaseTheme } from "metabase/embedding-sdk/theme";
 import { MetabaseReduxProvider } from "metabase/lib/redux";
 import { LocaleProvider } from "metabase/public/LocaleProvider";
 import { setOptions } from "metabase/redux/embed";
