+++ enterprise/frontend/src/embedding-sdk/lib/theme/index.ts
@@ -1,3 +1,1 @@
-export { DEFAULT_METABASE_COMPONENT_THEME } from "./default-component-theme";
 export { getEmbeddingThemeOverride } from "./get-embedding-theme";
-export { setGlobalEmbeddingColors } from "./embedding-color-palette";
