--- enterprise/frontend/src/embedding-sdk/types/utils.ts
@@ -1,4 +1,0 @@
-/** Makes every property in the object optional. */
-export type DeepPartial<T> = {
-  [K in keyof T]?: T[K] extends object ? DeepPartial<T[K]> : T[K];
-};
