+++ frontend/src/metabase/embedding-sdk/theme/get-embedding-chart-colors.ts
