+++ frontend/src/metabase/embedding-sdk/theme/index.ts
@@ -1,0 +1,3 @@
+export * from "./default-component-theme";
+// eslint-disable-next-line no-literal-metabase-strings -- file name
+export * from "./MetabaseTheme";
