--- frontend/src/metabase/embedding-sdk/types/fonts.ts
@@ -1,21 +1,0 @@
-export type MetabaseFontFamily =
-  | "Roboto"
-  | "Merriweather"
-  | "Open Sans"
-  | "Lato"
-  | "Noto Sans"
-  | "Roboto Slab"
-  | "Source Sans Pro"
-  | "Raleway"
-  | "Slabo 27px"
-  | "PT Sans"
-  | "Poppins"
-  | "PT Serif"
-  | "Roboto Mono"
-  | "Roboto Condensed"
-  | "Playfair Display"
-  | "Oswald"
-  | "Ubuntu"
-  | "Montserrat"
-  | "Lora"
-  | "Custom";
