+++ frontend/src/metabase/ui/theme.ts
@@ -1,7 +1,7 @@
 import type { MantineThemeOverride } from "@mantine/core";
 import { rem } from "@mantine/core";
 
-import { DEFAULT_METABASE_COMPONENT_THEME } from "embedding-sdk/lib/theme";
+import { DEFAULT_METABASE_COMPONENT_THEME } from "metabase/embedding-sdk/theme";
 
 import {
   getAccordionOverrides,
