+++ frontend/src/metabase/visualizations/visualizations/BarChart/BarChart.stories.tsx
@@ -5,7 +5,7 @@ import {
   VisualizationWrapper,
 } from "__support__/storybook";
 import { NumberColumn, StringColumn } from "__support__/visualizations";
-import type { MetabaseTheme } from "metabase/embedding-sdk/types/theme";
+import type { MetabaseTheme } from "metabase/embedding-sdk/theme";
 import { Box } from "metabase/ui";
 import { registerVisualization } from "metabase/visualizations";
 import Visualization from "metabase/visualizations/components/Visualization";
